+++ db/migrate/20170322001657_add_file_row_sampes_to_imports.rb
@@ -1,0 +1,5 @@
+class AddFileRowSampesToImports < ActiveRecord::Migration
+  def change
+    add_column :imports, :file_row_samples, :text
+  end
+end